+++ docs/docs/guides/response/config-pydantic.md
@@ -74,7 +74,7 @@ schema can be used to build another class with a custom config like:
 
 ```python hl_lines="10"
 from django.contrib.auth.models import User
-from ninja.orm import create_schema
+from ninja.schema import create_schema
 
 
 BaseUserSchema = create_schema(User)
