--- ninja/orm/__init__.py
@@ -1,4 +1,0 @@
-from ninja.orm.factory import create_schema
-from ninja.orm.metaclass import ModelSchema
-
-__all__ = ["create_schema", "ModelSchema"]
